feat: consolidate WorkLocation components into WorkLocationForm for improved structure and functionality
