+++ frontend/pages/community/configurations/work-location/create.tsx
@@ -3,7 +3,7 @@ import { useRouter } from "next/router";
 
 import ContentLayout from "~community/common/components/templates/ContentLayout/ContentLayout";
 import { useTranslator } from "~community/common/hooks/useTranslator";
-import CreateWorkLocation from "~community/configurations/components/organisms/CreateWorkLocation/CreateWorkLocation";
+import WorkLocationForm from "~community/configurations/components/organisms/WorkLocationForm/WorkLocationForm";
 
 const WorkLocationCreatePage: NextPage = () => {
   const router = useRouter();
@@ -17,7 +17,7 @@ const WorkLocationCreatePage: NextPage = () => {
       isBackButtonVisible
       onBackClick={() => router.back()}
     >
-      <CreateWorkLocation />
+      <WorkLocationForm />
     </ContentLayout>
   );
 };
